+++ src/main/scala/bouken/types/Pathing.scala
@@ -1,4 +1,4 @@
-package bouken
+package bouken.types
 
 import bouken.domain.{Place, Position, Route}
 import simulacrum._
