+++ src/test/scala/bouken/PathingTest.scala
@@ -1,6 +1,7 @@
 package bouken
 
 import bouken.domain._
+import bouken.types.MoveCosts
 import org.scalatest.{FreeSpec, Matchers}
 
 class PathingTest extends FreeSpec with Matchers {
@@ -8,7 +9,7 @@ class PathingTest extends FreeSpec with Matchers {
   import PathingTest._
 
   "Pathing" - {
-    import Pathing.ops._
+    import bouken.types.Pathing.ops._
     "isOutOfBounds" - {
       "is true when given a negative X Position" in {
         blankArea.isOutOfBounds(Position(-1, 0)) shouldBe true
@@ -119,7 +120,7 @@ class PathingTest extends FreeSpec with Matchers {
   }
 
   "Find Position" - {
-      import LocatePosition.ops._
+      import bouken.types.LocatePosition.ops._
       "Returns a place if one exists" in {
         val result = blankArea.find(Position(0, 0))
 
@@ -134,7 +135,7 @@ class PathingTest extends FreeSpec with Matchers {
   }
 
   "Suggest Route" - {
-    import Navigation.ops._
+    import bouken.types.Navigation.ops._
     "When a single route exists" - {
       val start = Position(0, 0)
       val finish = Position(0, 2)
